% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.design.R
\name{blocking.design}
\alias{blocking.design}
\title{Blocking Design}
\usage{
blocking.design(seed, num.per.slide = 8, num.array)
}
\arguments{
\item{seed}{an integer used to initialize a pseudorandom number generator.}

\item{num.per.slide}{number of arrays per array slide. It must be a multiple of 2. By default, \code{num.per.slide = 8} for Agilent microarrays.}

\item{num.array}{number of arrays. It must be a multiple of \code{num.per.slide}.}
}
\value{
a vector of array IDs in the order of assigning to samples that are assumed to be sorted by sample group of interest
As a result, the first half of the array IDs are assigned to sample group 1 and the second half of the array IDs are assigned to sample group 2.
}
\description{
Assign arrays to samples with blocking experimental design by array slide.
For example Agilent comes with an eight-plex design.
That is the number of arrays in an array slide is eight;
within each array slide, equal number of arrays are assigned to either sample group.
}
\examples{
blocking.design(seed = 1, num.per.slide = 8, num.array = 128)

}
\keyword{study.design}
