% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanc.intcv.R
\name{clanc.intcv}
\alias{clanc.intcv}
\title{Classification to Nearest Centroids  Classifier}
\usage{
clanc.intcv(kfold = 5, X, y, seed)
}
\arguments{
\item{kfold}{number of folds. By default, \code{kfold = 5}.}

\item{X}{dataset to be trained. This dataset must have rows as probes and columns as samples.}

\item{y}{a vector of sample group of each sample for the dataset to be trained.
It must have an equal length to the number of samples in \code{X}.}

\item{seed}{an integer used to initialize a pseudorandom number generator.}
}
\value{
a list of 4 elements:
\item{mc}{an internal misclassification error rate}
\item{time}{the processing time of performing internal validation with ClaNC}
\item{model}{a ClaNC classifier, resulted from \code{cv.fit}}
}
\description{
Build a Classification to Nearest Centroids classifier on the objective data.
}
\examples{
set.seed(101)
biological.effect <- estimate.biological.effect(uhdata = uhdata.pl)
ctrl.genes <- unique(rownames(uhdata.pl))[grep("NC", unique(rownames(uhdata.pl)))]
biological.effect.nc <- biological.effect[!rownames(biological.effect)
  \%in\% ctrl.genes, ]
group.id <- substr(colnames(biological.effect.nc), 7, 7)

biological.effect.train.ind <- colnames(biological.effect.nc)[c(sample(which(
  group.id == "E"), size = 64),
  sample(which(group.id == "V"), size = 64))]
biological.effect.nc.tr <- biological.effect.nc[, biological.effect.train.ind]

clanc.int <- clanc.intcv(X = biological.effect.nc.tr,
                     y = substr(colnames(biological.effect.nc.tr), 7, 7),
                     kfold = 5, seed = 1)

}
\references{
Alan R. Dabney, Author Notes.(2005) ClaNC: point-and-click software for classifying microarrays to nearest centroids,
https://academic.oup.com/bioinformatics/article/22/1/122/219377
}
\keyword{classification}
