% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{compute_log_lik}
\alias{compute_log_lik}
\title{Compute Negative Log Likelihood}
\usage{
compute_log_lik(prediction, actual)
}
\arguments{
\item{prediction}{Numeric vector same length as actual}

\item{actual}{Numeric vector same length as prediction}
}
\value{
Numeric vector length one
}
\description{
Bernoulli likelihood: x = 0, 1; f(x,theta) = (theta^x)*(1-theta)^(1-x)
Bernoulli log-likelihood: $$ ln(L) = sum_{i=1}^I sum_{t=1}^T  D_i^C(t)  *  ln(P_{i}^{C} (t)) + (1 - D_i^C(t)) * ln(1 - P_{i}^{C} (t)) $$
}
\examples{
compute_log_lik(runif(10), sample(c(1,0), 10, replace=TRUE))

}

