% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{compute_rmse}
\alias{compute_rmse}
\title{Compute Root Mean Squared Error}
\usage{
compute_rmse(prediction, actual)
}
\arguments{
\item{prediction}{Numeric vector same length as actual}

\item{actual}{Numeric vector same length as prediction}
}
\value{
Numeric vector length one
}
\description{
Compute Root Mean Squared Error
}
\examples{
compute_rmse(1:10, 1:10)
compute_rmse(1:10, c(1:9, 1))
compute_rmse(rnorm(10), rnorm(10))

}

