% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def_computeIters.R
\docType{class}
\name{computeIters-class}
\alias{computeIters-class}
\alias{plot,computeIters-method}
\alias{print,computeIters-method}
\alias{show,computeIters-method}
\alias{summary,computeIters-method}
\title{An S4 class to return the results of computing iterations needed}
\usage{
\S4method{summary}{computeIters}(object, digits = 3)

\S4method{print}{computeIters}(x, ...)

\S4method{show}{computeIters}(object)

\S4method{plot}{computeIters}(x, outcome_var = "Outcome",
  xlab = "Iterations", ylab = NULL)
}
\arguments{
\item{object}{S4 computeIters object

 @export}

\item{digits}{Optional numeric vector length one for how many significant digits to
print, default is 3.

 @export}

\item{x}{The result slot of an object created by \code{compute_iters}.}

\item{...}{ignored
@export}

\item{outcome_var}{Optional character vector.}

\item{xlab}{Optional character vector.}

\item{ylab}{Optional character vector.}
}
\value{
Returns a ggplot2 object.
}
\description{
An S4 class to return the results of computing iterations needed

Prints (cat()) summary of computeIters S4 object

Plots computeIters S4 object
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: 

\item \code{print}: An S4 method for printing a computeIters S4 object

\item \code{show}: An S4 method for showing a computeIters S4 object

\item \code{plot}: 
}}
\section{Slots}{

\describe{
\item{\code{call}}{Language from the call of the function.}

\item{\code{measure}}{Character vector length one with the measure used.}

\item{\code{result}}{c("src", "pcc") s3 classes from \code{sensitivity} package.}

\item{\code{plot_data}}{data.frame with two columns.}

\item{\code{timing}}{Numeric vector length one with the total elapsed time it took 
to execute.}

\item{\code{session}}{the results from calling \code{sessionInfo()} at end of
\code{\link{pc_sa}} function.}
}}

