% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounding.design.R
\name{confounding.design}
\alias{confounding.design}
\title{Confounding Design}
\usage{
confounding.design(seed, num.array, degree = "complete", rev.order = FALSE)
}
\arguments{
\item{seed}{an integer used to initialize a pseudorandom number generator.}

\item{num.array}{number of arrays.}

\item{degree}{level of confounding. It must be either "complete" or "partial"
for complete confounding design or partial confounding design, correspondingly.
By default, \code{degree = "complete"}.}

\item{rev.order}{whether the array-to-sample-group assignment should be flipped.
Originally, the first half arrays are designated to be assigned to sample group 1 (the endometrial sample group)
and the second half to sample group 2 (the ovarian sample group).
If the array-to-sample-group assignment is flipped (\code{rev.order = TRUE}),
the first half of the array IDs will be swapped with the second half of the array IDs.
By default, \code{rev.order = FALSE}.}
}
\value{
a vector of array IDs in the order of assigning to samples that are assumed to be sorted by sample group of interest
(first half of the samples belong to sample group 1 and second half to sample group 2).
As a result, the first half of the array IDs are assigned to sample group 1 and the second half of the array IDs are assigned to sample group 2.
}
\description{
Assign arrays to samples with confounding experimental design, intentionally assigning arrays to sample groups in the order of array collection.
Since we split the nonuniformly-handled training dataset in a particular way so that its earlier arrays were processed by one technician and the later arrays by the other,
assigning the earlier arrays to one sample group and the later arrays to another
results in confounding handling effects with the sample groups.
}
\examples{

# Completely confounding with reversed assignment
cc.rev.ind <- confounding.design(seed = 1, num.array = 128,
                             degree = "complete", rev.order = FALSE)

# Partially confounding
pc.ind <- confounding.design(seed = 1, num.array = 128,
                             degree = "partial")
}
\keyword{study.design}
