% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.storage.R
\name{create.storage}
\alias{create.storage}
\title{Create Storage for Output}
\usage{
create.storage(
  design.list = c("CC+", "CC-", "PC+", "PC-"),
  class.list = c("PAM", "LASSO"),
  norm.list = c("NN", "QN"),
  validating.sets = c("test", "gold.standard")
)
}
\arguments{
\item{design.list}{List of design methods to be applied in \code{precision.simulate}.}

\item{class.list}{List of classification methods to be applied in \code{precision.simulate}}

\item{norm.list}{List of normalization methods to be applied in \code{precision.simulate}}

\item{validating.sets}{Types of validation methods.}
}
\value{
A list whose size is determined by the input parameters.
}
\description{
During the implement of \code{precision.simulate} function, according to the length of \code{design.list},
\code{class.list}, and \code{norm.list}, to create a storage list to store the output from \code{precision.simulate}.
}
\details{
create.storage
}
\examples{
create.storage(design.list = c("CC+", "CC-", "PC+", "PC-"),
               class.list = c("PAM", "LASSO"),
               norm.list = c("NN", "QN"),
               validating.sets = c("test", "gold.standard"))


}
\keyword{storage}
