% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{create_loss_func}
\alias{create_loss_func}
\title{Pick one of the loss functions from the package.}
\usage{
create_loss_func(loss)
}
\arguments{
\item{loss}{Character vector length one specifying which loss function you
want.}
}
\value{
Returns a function that has two arguments \code{prediction} (Numeric
  vector same length as actual), and \code{actual} (Numeric vector same
  length as prediction), and returns a Numeric vector length one.
}
\description{
Pick one of the loss functions from the package.
}
\examples{
create_loss_func("log_lik")

}

