% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.biological.effect.R
\name{estimate.biological.effect}
\alias{estimate.biological.effect}
\title{Estimated Sample Effects}
\usage{
estimate.biological.effect(uhdata)
}
\arguments{
\item{uhdata}{the uniformly-handled dataset.
The dataset must have rows as probes and columns as samples.}
}
\value{
an estimation of the sample effects
}
\description{
Estimate biological effects of a sample from the uniformly-handled dataset.
}
\examples{
biological.effect <- estimate.biological.effect(uhdata = uhdata.pl)

}
\keyword{data.setup}
