% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.precision.error.R
\name{extract.precision.error}
\alias{extract.precision.error}
\title{Extracting errors from PRECISION (both non-FLEX and FLEX) output}
\usage{
extract.precision.error(precision.obj.path = NULL, precision.obj.name)
}
\arguments{
\item{precision.obj.path}{the path and document name for the PRECISION object. By default, \code{precision.obj.path = NULL} indicates that the PRECISION object is already existed in the current global environment.}

\item{precision.obj.name}{a string for the name of the PRECISION R object.}
}
\value{
a list of lists containing all the errors from the PRECISION object with the same names as the object
}
\description{
Extract all errors in the \code{error_store} list from PRECISION output produced from either \code{precision.simulate.flex}; reduce the PRECISION output for plotting.
}
\examples{
\dontrun{
## Load + extract errors from a PRECISION object
precision.error.df <- extract.precision.error(precision.obj.path =
  "mypath/precision.result.Rdata",
  precision.obj.name = "precision.result")
  
## Load + extract errors from a PRECISION FLEX object
precision.flex.error.df <- extract.precision.error(precision.obj.path =
  "mypath/precision.flex.result.Rdata",
  precision.obj.name = "precision.flex.result")

## Extract errors from a PRECISION FLEX object previously existed in global env
load("mypath/precision.a.Rdata") # load object named "precision.a" in global env
precision.flex.error.df <- extract.precision.error(precision.obj.name = "precision.flex.result")
}

}
\keyword{result.manage}
