% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{Fill_NA}
\alias{Fill_NA}
\title{Fill NA values}
\usage{
Fill_NA(data, replace = 0)
}
\arguments{
\item{data}{A vector or data frame}

\item{replace}{character. If there exists \code{'MEAN'} in given argument then replace \code{NA} values with \code{\emph{mean}}.
if \code{'MEDIAN'} then replace with \code{\emph{median}}, otherwise replace with 0.}
}
\value{
Returns \code{NA} filled vector or data frame
}
\description{
The function \code{Fill_NA()} is used to fill \code{NA} values with the same data type values.
It works on vectors & Data Frames with respect to character, factor, numeric,integer data types.
If any other data types is used this function can't be apply.
}
\examples{
v1 <- c(1,2,3,NA, 5.2)
v2 <- c(NA,'a','a','b','d')
v3 <- c(3+2i,4,6,NA,1)
df <- data.frame(v1,v2,v3);df
str(df)
Fill_NA(df) # example of \code{data.frame}
Fill_NA(df,  replace = 'MEDIAN')  example with data.frame
Fill_NA(v1, replace = 'MEAN')  example with vector

}
\seealso{
\link{Missing_C} for missing count, \link{Missing_P} getting missing percentage
}
\author{
Naveen Kumar M.Sc., \emph{Email}: \email{info@saraswathianalytics.com} OR \emph{WhatsApp}: \href{https://wa.me/918688896472}{Click Here}
}
