% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.pbset.R
\name{limma.pbset}
\alias{limma.pbset}
\title{Differential expression analysis of probe-set data}
\usage{
limma.pbset(data, group.id, group.id.level = c("E", "V"), pbset.id = NULL)
}
\arguments{
\item{data}{dataset to be analyzed.
The dataset must have rows as unique probe-sets and columns as samples.}

\item{group.id}{a vector of sample-group labels for each sample of the dataset.
It must be a 2-level non-numeric factor vector.}

\item{group.id.level}{a vector of sample-group label level.
It must have two and only two elements and the first element is the reference.
By default, \code{group.id.level = c("E", "V")}.
That is in our study, we compare endometrial tumor samples to
ovarian tumor samples, with endometrial as our reference.}

\item{pbset.id}{a vector of unique probe-set names.
By default, \code{pbset.id = NULL} for it to be the row names of the dataset.}
}
\value{
a data frame with differential expression analysis results,
group means and group standard deviations, for each unique probe-set.
}
\description{
Perform two-group differential expression analysis using "limma".
}
\examples{
uhdata.psl <- med.sum.pbset(data = uhdata.pl,
                            num.per.unipbset = 10)
ctrl.genes <- unique(rownames(uhdata.pl))[grep("NC", unique(rownames(uhdata.pl)))]

uhdata.psl.nc <- uhdata.psl[!rownames(uhdata.psl) \%in\% ctrl.genes, ]

group.id <- substr(colnames(uhdata.psl.nc), 7, 7)
group.id.level <- levels(as.factor(group.id))

limma.fit.uhdata<- limma.pbset(data = uhdata.psl.nc,
                               group.id = group.id,
                               group.id.level = group.id.level)
                               table(limma.fit.uhdata$P.Value < 0.01,
                               dnn = "DE genes")

}
\references{
Ritchie M., Phipson B., Wu D., Hu Y., Law C., Shi W. and Smyth G. (2015).
"limma powers differential expression analyses for RNA-sequencing and microarray studies."
Nucleic Acids Research, 43(7), pp. e47.
}
\keyword{DEA}
