% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.sum.pbset.R
\name{med.sum.pbset}
\alias{med.sum.pbset}
\title{Probe-set median summarization}
\usage{
med.sum.pbset(data, pbset.id = NULL, num.per.unipbset = 10)
}
\arguments{
\item{data}{dataset to be summarized.
The dataset must have rows as probes and columns as samples.
It must be a data matrix with the same number of probes per unique probe-set.
If it is already on the probe-set level, no manipulation will be done.}

\item{pbset.id}{a vector of unique probe-set names.
If it is not specified, then by default it is set to be the unique probe names of the data.}

\item{num.per.unipbset}{number of probes for each unique probe-set.
By default, \code{num.per.unipbset = 10}.}
}
\value{
probe-set median summarized data
}
\description{
Summarize probe-set using median of each unique probe and
only takes in data matrix with the same number of probes per unique probe-set.
}
\examples{
\dontrun{
uhdata.psl <- med.sum.pbset(data = uhdata.pl,
                            num.per.unipbset = 10)
}

}
\keyword{preprocess}
