% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{Missing_P}
\alias{Missing_P}
\title{Missing Percentage}
\usage{
Missing_P(x)
}
\arguments{
\item{x}{A vector or data frame}
}
\value{
The missing percentage of input
}
\description{
The function \code{Missing_P} returns missing percentage of the given vector (or) data frame.
It's working on vectors & Data Frames.
}
\examples{
v1 <- c('A','B','C',NA,'D',NA)
v2 <- c(NA,1,2,3,2,6)
v3 <- c(3+2i,8,6,NA,1,NA)
v4 <- c(3,8,6,1,1,7)
df <- data.frame(v1,v2,v3,v4);df
Missing_P(v1) # passing vector
Missing_P(df) # passing dataframe
}
\seealso{
\link{Fill_NA} for filling \code{NA} values, \link{Mode} for mode, \link{Missing_C} for missing count
}
\author{
Naveen Kumar M.Sc., \emph{Email}: \email{info@saraswathianalytics.com} OR \emph{WhatsApp}: \href{https://wa.me/918688896472}{Click Here}
}
