% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(data)
}
\arguments{
\item{data}{A vector or data frame}
}
\value{
The mode of the input
}
\description{
The function \code{Mode} returns mode of the given vector (or) data frame along with same data type values.
It's working on vectors & Data Frames.
}
\examples{
v1 <- c("B","A","A","B","A")
v2 <- c(NA,1,2,3,2)
v3 <- c(3+2i,8,6,NA,1)
df <- data.frame(v1,v2,v3);df
Mode(v1) # passing vector
Mode(df) # passing dataframe
}
\seealso{
\link{Fill_NA} for filling \code{NA} values, \link{Missing_P} for missing percentage and \link{Missing_C} for missing count
}
\author{
Naveen Kumar M.Sc., \emph{Email}: \email{info@saraswathianalytics.com} OR \emph{WhatsApp}: \href{https://wa.me/918688896472}{Click Here}
}
