% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def_pcSA.R
\docType{class}
\name{pcSA-class}
\alias{pcSA-class}
\alias{plot,pcSA-method}
\alias{print,pcSA-method}
\alias{show,pcSA-method}
\title{An S4 class to return the results of sensitivity analyses}
\usage{
\S4method{print}{pcSA}(x, ...)

\S4method{show}{pcSA}(object)

\S4method{plot}{pcSA}(x, outcome_var = "Outcome", xlab = "Variable",
  ylab = "Standardized Rank Regression Coefficient")
}
\arguments{
\item{x}{The result slot of an object created by \code{pc_sa}.}

\item{...}{ignored
@export}

\item{object}{S4 pcSA object

 @export}

\item{outcome_var}{Optional character vector for labeling the outcome variable
in the plot. Default is "Outcome".}

\item{xlab}{Optional character vector for labeling the variables.}

\item{ylab}{Optional character vector. Default is "Partial Rank Correlation
Coefficient". Could also be "Partial Correlation Coefficient", "Standardized
Rank Regression Coefficient", and "Standardized Regression Coefficient".}
}
\value{
Returns a ggplot2 plot.
}
\description{
An S4 class to return the results of sensitivity analyses

Show

Plots pcSA S4 object, a Partial Correlation Analysis of a Simulation Model
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: An S4 method for printing a pcSA S4 object

\item \code{show}: An S4 method for showing a pcSA S4 object

\item \code{plot}: This is function of the \strong{eat} package. \code{pc_sa} conducts a a 
partial correlation analysis.
}}
\section{Slots}{

\describe{
\item{\code{call}}{Language from the call of the function.}

\item{\code{input_set}}{data.frame of input set used}

\item{\code{sims}}{numeric vector with outcome of simulating model with input_set}

\item{\code{result}}{c("src", "pcc") s3 classes from \code{sensitivity} package.}

\item{\code{r_squared}}{Numeric vector length one.}

\item{\code{rmse}}{Numeric vector length one.}

\item{\code{timing}}{Numeric vector length one with the total elapsed time it took to execute.}

\item{\code{session}}{the results from calling \code{sessionInfo()} at end of
\code{\link{pc_sa}} function.}
}}
\examples{
fake_abm <- function(params, out) {
  x1 <- params[1]
  x2 <- params[2]
  if (out=="sq") return(x1^2 + x2 + rnorm(1, 0))
  if (out=="ident") return(x1 + x2 + rnorm(1, 0))
}
inputs <- lapply(list(param1 = NA, param2 = NA), 
                 function(x) list(random_function = "qunif",
                                  ARGS = list(min = 0, max = 1)))
s <- pc_sa(fake_abm, inputs, "sq")
plot(s)

}
\references{
J. C. Thiele, W. Kurth, V. Grimm, Facilitating Parameter 
 Estimation and Sensitivity Analysis of Agent-Based Models: A Cookbook Using 
 NetLogo and R. Journal of Artificial Societies and Social Simulation. 17, 11
 (2014).
}

