% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per.unipbset.truncate.R
\name{per.unipbset.truncate}
\alias{per.unipbset.truncate}
\title{Probe-level data truncation to a fixed number of probes per unique probe-set}
\usage{
per.unipbset.truncate(data, pbset.id = NULL, num.per.unipbset = 10)
}
\arguments{
\item{data}{probe-level dataset. The dataset must have rows as probes and columns as samples.}

\item{pbset.id}{a vector of unique probe-set names.
By default, \code{pbset.id = NULL} for it to be the row names of the dataset.}

\item{num.per.unipbset}{number of probes for each unique probe-set to be truncated to.
By default, \code{num.per.unipbset = 10}.}
}
\value{
truncated probe-level data
}
\description{
Truncate probe-level dataset so that it has a fixed number of probes per unique probe-set.
We are safe to do so if the variation among replicates for the same probe is small.
}
\examples{
uhdata.pl.p5 <- per.unipbset.truncate(data = uhdata.pl,
num.per.unipbset = 5)
}
\keyword{data.setup}
