% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.precision.R
\name{plot.precision}
\alias{plot.precision}
\title{Plot misclassification error rates from PRECISION (both non-FLEX and FLEX) output}
\usage{
\method{plot}{precision}(
  data,
  mytitle = "PRECISION results (misclass. error rates)",
  design.order = NULL,
  ylim = c(0, 0.5),
  iflex = TRUE
)
}
\arguments{
\item{data}{a resulted PRECISION output from \code{extract.precision.error}.}

\item{mytitle}{the title of the plot.}

\item{design.order}{indices in the order of what user would like to plot from the resulted output from \code{extract.precision.error}. By default, every simulation is plotted.}

\item{ylim}{y limit. By default, \code{ylim = c(0, 0.5)}.}

\item{iflex}{whether data if a PRECISION output from \code{precision.simulate} or \code{precision.simulate.flex}. By default, \code{iflex = TRUE} indicating that the data are from \code{precision.simulate.flex}.}
}
\value{
a boxplot of misclassification error rates
}
\description{
Plot error rates from PRECISION output that is being extracted by \code{extract.precision.error}
}
\examples{
\dontrun{
## PRECISION output
plot.precision(data = precision.results.err.df,
  design.order = 1:10,
  mytitle = "My PRECISION results", iflex = FALSE)

## PRECISION FLEX output
plot.precision(data = precision.results.flex.err.df,
  design.order = 1:10,
  mytitle = "My PRECISION FLEX results")

# exclude ext.uh
plot.precision(data = precision.results.flex.err.df[!names(precision.results.flex.err.df) \%in\% "ext.uh"],
  design.order = 1:10,
  mytitle = "My PRECISION FLEX results")
}

}
\keyword{result.manage}
