% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranfor.intcv.R
\name{ranfor.intcv}
\alias{ranfor.intcv}
\title{Random Forest  Classifier}
\usage{
ranfor.intcv(kfold = 5, X, y, seed)
}
\arguments{
\item{kfold}{number of folds. By default, \code{kfold = 5}.}

\item{X}{dataset to be trained. This dataset must have rows as probes and columns as samples.}

\item{y}{a vector of sample group of each sample for the dataset to be trained.
It must have an equal length to the number of samples in \code{X}.}

\item{seed}{an integer used to initialize a pseudorandom number generator.}
}
\value{
a list of 4 elements:
\item{mc}{an internal misclassification error rate}
\item{time}{the processing time of performing internal validation with LASSO}
\item{model}{a LASSO classifier, resulted from \code{cv.fit}}
}
\description{
Build a random forest classifier using internal cross validation to choose the turning parameter, with a 5-fold cross validation as default.
}
\examples{
set.seed(101)
biological.effect <- estimate.biological.effect(uhdata = uhdata.pl)
ctrl.genes <- unique(rownames(uhdata.pl))[grep("NC", unique(rownames(uhdata.pl)))]
biological.effect.nc <- biological.effect[!rownames(biological.effect)
  \%in\% ctrl.genes, ]
group.id <- substr(colnames(biological.effect.nc), 7, 7)

biological.effect.train.ind <- colnames(biological.effect.nc)[c(sample(which(
  group.id == "E"), size = 64),
  sample(which(group.id == "V"), size = 64))]
biological.effect.nc.tr <- biological.effect.nc[, biological.effect.train.ind]

ranfor.int <- ranfor.intcv(X = biological.effect.nc.tr,
                         y = substr(colnames(biological.effect.nc.tr), 7, 7),
                         kfold = 5, seed = 1)

}
\references{
https://cran.r-project.org/web/packages/e1071/index.html
}
\keyword{classification}
