% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\docType{package}
\name{sa}
\alias{sa}
\alias{sa-package}
\title{sa: A Package for Sensitivity Analysis for Complex Computational Models}
\description{
we have settled on one or a few with good performance:
\enumerate{
\item \code{\link{create_sample}}
\item \code{\link{sobol_sa}}
\item \code{\link{pc_sa}}
}
}

