% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.design.R
\name{stratification.design}
\alias{stratification.design}
\title{Stratification Design}
\usage{
stratification.design(seed, num.array, batch.id)
}
\arguments{
\item{seed}{an integer used to initialize a pseudorandom number generator.}

\item{num.array}{number of arrays.}

\item{batch.id}{a list of array indices grouped by batches when data were profiled.
The length of the list must be equal to the number of batches in the data;
the number of array indices must be the same as the number of samples.}
}
\value{
a vector of array IDs in the order of assigning to samples that are
assumed to be sorted by sample group of interest
As a result, the first half of the array IDs are assigned to group 1
and the second half of the array IDs are assigned to group 2.
}
\description{
Assign arrays to samples with stratification,
a study design assigning arrays in each batch to each sample group proportionally.
}
\examples{
batch.id <- list(1:40, 41:64, (129:160) - 64, (161:192) - 64)
str.ind <- stratification.design(seed = 1, num.array = 128,
                                 batch.id = batch.id)
}
\keyword{study.design}
