% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.classifier.funcs.R
\name{switch.classifier.funcs}
\alias{switch.classifier.funcs}
\title{Switch classfication functions}
\usage{
switch.classifier.funcs(
  class.list = c("PAM", "LASSO"),
  class.funcs = NULL,
  pred.funcs = NULL
)
}
\arguments{
\item{class.list}{A list of classification method name to be transformed. The build-in methods are "PAM" and "LASSO".}

\item{class.funcs}{Other classfication functions that users can create by themselves.}

\item{pred.funcs}{Other prediction funcstions that users can create by themselves.}
}
\value{
A list containing \code{build.funcs}, which means the fitting function names, and \code{pred.funcs},
which means the prediction function names. The fitting functions is to be fitted on the training set and prediction
function is to be used on the test set.
}
\description{
Transform classfication method names into a fitting function name and a prediction function name,
for using in the classfication procedure
}
\details{
switch.classifier.funcs
}
\examples{
switch.classifier.funcs.class(class.list =  c("PAM", "LASSO"),
                              class.funcs = NULL, pred.funcs = NULL)


}
\keyword{names}
\keyword{switch}
