% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.norm.funcs.flex.R
\name{switch.norm.funcs.flex}
\alias{switch.norm.funcs.flex}
\title{Switch normalization funcsions in a flexible way}
\usage{
switch.norm.funcs.flex(norm.list = c("NN", "QN"), norm.funcs = NULL)
}
\arguments{
\item{norm.list}{Switch all the build-in normalization methods into function names, including "NN", "MN", "QN", and "VSN".}

\item{norm.funcs}{New functions that user can create by themselves.}
}
\value{
A list that transforms the normalization method name into function name.
}
\description{
Transform the normalization method names into functions names, for running in the normalization procedure
}
\details{
switch.norm.funcs.flex
}
\examples{
switch.norm.funcs.flex(norm.list = c("NN", "MN", "QN"), norm.funcs = NULL)

}
\keyword{names}
\keyword{switch}
