% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.ext.err.func.R
\name{tabulate.ext.err.func}
\alias{tabulate.ext.err.func}
\title{Tabulate.ext.err.func}
\usage{
tabulate.ext.err.func(pred.obj, obs.grp)
}
\arguments{
\item{pred.obj}{predicted group classes}

\item{obs.grp}{actual group classes}
}
\value{
error rate of predicted classification
}
\description{
An assistant function used to calculate the error rate of classification.
}
\examples{
pred.obj = c("V", "V", "V", "V", "E", "V", "E", "E", "E", "E")
obs.grp = c("V", "V", "V", "V", "V", "E", "E", "E", "E", "E")
tabulate.ext.err.func(pred.obj, obs.grp)

}
\keyword{classification}
