% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{X_Mas}
\alias{X_Mas}
\title{making X_Mas tree}
\usage{
X_Mas(locator = "topright")
}
\arguments{
\item{locator}{A \code{character}, if \code{locator = 'locator'} Reads the position of the graphics cursor when the (first) mouse button is pressed or
location may also be specified by setting x to a single keyword from the list \code{"bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright",
 "right" and "center"}. This places the legend on the inside of the plot frame at the given location.}
}
\value{
Return Christmas tree
}
\description{
The function \code{X_Mas} returns Christmas tree.
}
\seealso{
\link{legend} for locating legend, \link{locator} for position of the graphics cursor.
}
\author{
Naveen Kumar M.Sc., \emph{Email}: \email{info@saraswathianalytics.com} OR \emph{WhatsApp}: \href{https://wa.me/918688896472}{Click Here}
}
